+++ jest.config.js
@@ -1,6 +1,9 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  clearMocks: true,
+  roots: ['<rootDir>/tests'],
+  testMatch: ['**/*.test.ts'],
   moduleNameMapper: {
     '^src/(.*)$': '<rootDir>/src/$1'
   }
